Make required renamings and update pnpm-lock.yaml
