+++ timetable/extractor.py
@@ -4,9 +4,9 @@ from dataclasses import dataclass
 from bs4 import BeautifulSoup
 from bs4.element import Tag, NavigableString, PageElement
 
-from utils.levenshtein_distance import get_minimum_distant_code
+from utils.minimum_distant_code import get_minimum_distant_code
 
-course_code_map = json.load('course_data.json')
+course_code_map = json.load(open('timetable/course_data.json'))
 
 with open("full_location.json") as data_file:
     full_locations = json.load(data_file)
